contrato360/pyproject.toml
[tool.poetry]
name = "contrato360"
version = "0.1.0"
description = ""
authors = ["João Azevedo <joao.nepomuceno@bndes.gov.br>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.35.0"
openai = "^1.7.2"
langchain = "^0.2.5"
langchain-openai = "^0.1.9"
langchain-community = "^0.2.6"
pysqlite3-binary = "^0.5.3"
python-dotenv = "^1.0.1"
plotly = "^5.23.0"


[tool.poetry.group.dev.dependencies]
jupyterlab="^3.2.0"
pyodbc = "^5.1.0"
ipykernel = "^6.29.4"
pandas = "~2.1.4"
pymupdf = "^1.24.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
